Move exceptions_builder into provider
